implemented opencv stream (sorta)

+++ src/cameraSocket.cpp
@@ -37,13 +37,13 @@ bool CameraSocket::establishConnection()
 }
 
 
-int CameraSocket::send(const unsigned char data[], int bytesToSend)
+int CameraSocket::send(const unsigned char data[], long bytesToSend)
 {
 
     return socket_.send_to(boost::asio::buffer(data, bytesToSend), serverEndpoint_);
 }
 
-int CameraSocket::receive(unsigned char dataBuf[], int size)
+int CameraSocket::receive(unsigned char dataBuf[], long size)
 {
     return socket_.receive_from(boost::asio::buffer(dataBuf, size), serverEndpoint_);
 }
