+++ src/serverSocket.hpp
@@ -7,8 +7,8 @@ class ServerSocket
 public:
     ServerSocket(std::string networkConfigPath);
     ~ServerSocket();
-    int send(const unsigned char data[], int bytesToSend);
-    int receive(unsigned char dataBuf[], int size);
+    int send(const unsigned char data[], long bytesToSend);
+    int receive(unsigned char dataBuf[], long size);
 
 private:
     boost::asio::io_service ioService_;
